server/requirements.txt
flask-restful
flask-sqlalchemy
mysqlclient